+++ tests/integration/solidity_contracts/conftest.py
@@ -94,7 +94,12 @@ def deploy_solidity_contract(kakarot, get_solidity_contract):
         )
         with traceit.context(contract_name):
             tx = await kakarot.eth_send_transaction(
-                to=0, gas_limit=1_000_000, gas_price=0, value=0, data=deploy_bytecode
+                origin=int(caller_eoa.address, 16),
+                to=0,
+                gas_limit=1_000_000,
+                gas_price=0,
+                value=0,
+                data=deploy_bytecode,
             ).execute(caller_address=caller_eoa.starknet_address)
 
         deploy_event = [